Your commit message

+++ src/main/java/com/kgisl/sb1/entity/Person.java
@@ -2,6 +2,6 @@ package com.kgisl.sb1.entity;
 
 import org.springframework.data.annotation.Id;
 
-public record Person(@Id Long id,String uname, String email) {
+public record Person(@Id int id,String uname, String email) {
     
 }
